requirements.txt
tensorflow==2.10.0
tensorflow-probability==0.17.0
torch==2.7.1
torchvision==0.22.1
scikit-learn==1.7.0
laplace-torch==0.2.2.2
netcal==1.3.6
matplotlib==3.7.5